+++ Reach/src/utopia/reach/container/multi/ContextualViewContainerFactory.scala
@@ -39,13 +39,15 @@ trait ContextualViewContainerFactory[+N, TopN, +Container <: ReachComponent, -To
 	  * @param pointer Pointer whose value is reflected in the container contents
 	  * @param contentFactory A factory used for constructing the individual components
 	  * @param construct A function that accepts
-	  *                  1) an initialized component-creation factory and
-	  *                  2) a pointer that contains the value to display on that component,
+	  *                     1. An initialized component-creation factory
+	  *                     1. A pointer that contains the value to display on that component
+	  *                     1. Index of this component
+	  *
 	  *                  and yields a new component
 	  * @tparam A Type of values displayed on individual components
 	  * @tparam F Type of component factories to use
 	  * @return A new container
 	  */
-	def mapPointer[A, F](pointer: Changing[Seq[A]], contentFactory: Ccff[N, F])(construct: (F, Changing[A]) => TopC) =
-		_mapPointer(pointer) { p => Open.withContext(context)(contentFactory) { construct(_, p) } }
+	def mapPointer[A, F](pointer: Changing[Seq[A]], contentFactory: Ccff[N, F])(construct: (F, Changing[A], Int) => TopC) =
+		_mapPointer(pointer) { (p, i) => Open.withContext(context)(contentFactory) { construct(_, p, i) } }
 }